Reverted failed fix and dropped `long double` test
`long double` has some bit representation complexity as it is
an 80-bit value padded to 128-bit and the copy constructor
doesn't copy the last 48 bits, so we have dropped
`long double` tests from test suite.

You should test `long double` literals manually and confirm
that the first 10 bytes are the same as the reference
implementation.